+++ src/test/java/numble/bankingapi/banking/domain/AccountServiceTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import numble.bankingapi.banking.exception.NotNegativeMoneyException;
-import numble.bankingapi.banking.tobe.ToBeAccountService;
 import numble.bankingapi.fake.FakeAccountHistoryRepository;
 import numble.bankingapi.fake.FakeAccountRepository;
 import numble.bankingapi.fake.FakeMemberRepository;
@@ -22,7 +21,7 @@ class AccountServiceTest {
 	private final FakeMemberRepository memberRepository = new FakeMemberRepository();
 	private final FakeAccountRepository accountRepository = new FakeAccountRepository();
 	private final FakeAccountHistoryRepository accountHistoryRepository = new FakeAccountHistoryRepository();
-	private ToBeAccountService accountService;
+	private AccountService accountService;
 	Account 사용자_계좌;
 	Account 상대방_계좌;
 	Member 사용자;
@@ -31,7 +30,7 @@ class AccountServiceTest {
 
 	@BeforeEach
 	void setUp() {
-		accountService = new ToBeAccountService(accountRepository, accountHistoryRepository);
+		accountService = new AccountService(accountRepository, accountHistoryRepository);
 		사용자 = memberRepository.save(new Member("member@email.com", "friend1", "password"));
 		사용자의_친구1 = memberRepository.save(new Member("member_friend1@email.com", "friend2", "password"));
 		사용자의_친구2 = memberRepository.save(new Member("member_friend2@email.com", "friend3", "password"));
